Break up `STTextViewController`
